backend/requirements.txt
fastapi
uvicorn
kafka-python
websockets
pandas
numpy
redis
